Static cast the streamoff to a size_t before allocating a new char buffer to avoid data loss warnings under some compilers

+++ src/core/testing/plg_panel_manager_test/test_asset_presentation_provider.cpp
@@ -35,7 +35,7 @@ ThumbnailData TestAssetPresentationProvider::getBinaryDataFromFile( const char *
 		return nullptr;
 	}
 	
-	char * buffer = new char[ length ];
+	char * buffer = new char[ static_cast< size_t >( length ) ];
 	input.read( buffer, length );
 	input.close();
